% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allCocktails}
\alias{allCocktails}
\title{The most updated list of cocktails and informations obtain from TheCocktailDB API}
\format{A data.frame with 402 rows and 53 columns}
\source{
\url{https://www.thecocktaildb.com/api.php}
}
\usage{
allCocktails
}
\description{
The data.frame includes all cocktails' ID, names and all other information available from the API
The API currently contains information for 402 cocktails
}
\keyword{datasets}
