% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CocktailSummary.R
\name{CocktailSummary}
\alias{CocktailSummary}
\title{Obtain cocktail summaries}
\usage{
CocktailSummary(summaryOption = "FirstLetter")
}
\arguments{
\item{summaryOption}{parameter allows user to choose from four different summary types. The
valid summary types are following: FirstLetter, IsAlcoholic, Category and Ingredients.
The parameter is defaulted to FirstLetter.
FirstLetter: provides summary over number of cocktails starts with 26 letters
IsAlcoholic: provides summary over number of cocktails by alcoholic type
Category: provides summary over number of cocktails by category
Ingredients: collects all ingredients used in all cocktails and returns 20 most
common ingredients}
}
\description{
This function allows user to obtain summaries of cocktails based on the
local data.frame contains the latest list of cocktails from pulled from the API
}
\examples{
CocktailSummary("Ingredients")
CocktailSummary("Category")
}
