% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllCocktails.R
\name{GetAllCocktails}
\alias{GetAllCocktails}
\title{Obtain all available cocktail information from the API}
\usage{
GetAllCocktails(precise = TRUE)
}
\arguments{
\item{precise}{parameter precise ask if the user want the result in precise format, if
set to be TRUE, the return information will only include drink id, name, category, type
glass used and date last modified
English instructions}
}
\description{
This function allows user to pull a complete list of cocktails from
TheCocktailDB API with all information available on the API site. The user get to
choose if data returns precise information. Please be aware that running this function
may take long time due to amount of information it's pulling
}
\examples{
GetAllCocktails()
GetAllCocktails(precise = TRUE)
}
