% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCocktailInfo.R
\name{GetCocktailInfo}
\alias{GetCocktailInfo}
\title{Search for cocktail information based on cocktail IDs}
\usage{
GetCocktailInfo(cocktailID = 11007, precise = FALSE)
}
\arguments{
\item{cocktailID}{asks the user to key in the cocktail ID, the ID list can be obtained
by the GetCocktailList function. Default value of ID is set to "11007" (margrita)}

\item{precise}{parameter precise ask if the user want the result in precise format, if
set to be TRUE, the return information will only include drink id, name, tag, type and
English instructions}
}
\description{
This function allows user to search for cocktail information based on
cocktail IDs. User is able to choose result format in either precise version or complete
version
}
\examples{
GetCocktailInfo(15266, precise = FALSE)
GetCocktailInfo(14167, precise = TRUE)
}
