% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInstruction.R
\name{GetInstruction}
\alias{GetInstruction}
\title{Obtain instruction for certain cocktails}
\usage{
GetInstruction(id = "15266", language = "EN", precise = TRUE)
}
\arguments{
\item{id}{asks user to enter cocktail IDs to search, the cocktail IDs are numberical IDs
obtained from the API and is also available in the package's included data.frame. Do a
c() command to search for multiple cocktails at a time}

\item{language}{This parameter allows user to choose the instruction language, allowed entries
include English (EN), Spanish (ES), German (DE), French (FR) and Chinese (ZH). The parameter is
defaulted to English (EN)}

\item{precise}{parameter precise ask if the user want the result in precise format, if
set to be TRUE, the return information will only include drink id, name, category, type
glass used and instruction}
}
\description{
This function allows user to obtain instruction for certain cocktails
based on the local data.frame contains the latest list of cocktails from pulled from
the API. User can do search for multiple cocktails at a time, and the instruction returned
may be in other languages as well
}
\examples{
GetInstruction(language = "DE")
GetInstruction(id = "15266", language = "EN")
GetInstruction(id = c("17840", "15266"))
}
